+++ src/include/node.h
@@ -9,9 +9,9 @@ class Node
 {
 public:
     const unsigned int id;
-    float x, y, z;
-    const float original_x, original_y, original_z;
+    double x, y, z;
+    const double original_x, original_y, original_z;
     std::vector<std::unique_ptr<Dof>> dofs;
     std::vector<Element*> connected_elements;
-    Node(unsigned int global_id, float x_, float y_, float z_ = 0);
+    Node(unsigned int global_id, double x_, double y_, double z_ = 0);
 };
